Refactor and add ::openvpn::server::dynamic define
This define allows simple OpenVPN server setups based on a template file and
keys and certificates used by Puppet, or ones generate with an external CA such
as Easy-RSA 3. This approach makes management of multiple servers significantly
easier than before, when the configuration files were always static and the keys
and certificates had to be embedded into the file. Note that as a significant
amount of refactoring was done while adding the define, so there is some
potential for breakages; this is reflected by the module version number.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>