Allow use of the new openvpn apt repositories
Previously there was only one official openvpn apt repository which had only
stable releases. Now there are several, and this commit adds support for using
all of them. Note that this commit removes the $use_latest_release parameter and
if it is set, the puppet run will fail with a clarifying error message.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>